+++ recipes/comboboxes/combobox_multi_select/combobox_multi_select.vue
@@ -11,6 +11,7 @@
     :visually-hidden-close="visuallyHiddenClose"
     content-width="anchor"
     :append-to="appendTo"
+    :transition="transition"
     @select="onComboboxSelect"
   >
     <template #input="{ onInput }">
@@ -20,7 +21,7 @@
       >
         <span
           ref="chipsWrapper"
-          class="d-ps-absolute d-mx2"
+          class="d-ps-absolute d-mx2 d-pl1"
         >
           <dt-chip
             v-for="item in selectedItems"
@@ -291,6 +292,15 @@ export default {
             (appendTo instanceof HTMLElement);
       },
     },
+
+    /**
+     * Named transition when the content display is toggled.
+     * @see DtLazyShow
+     */
+    transition: {
+      type: String,
+      default: 'fade',
+    },
   },
 
   emits: [
@@ -440,7 +450,7 @@ export default {
   methods: {
     onChipRemove (item) {
       this.$emit('remove', item);
-      this.$refs.input.focus();
+      this.$refs.input?.focus();
     },
 
     onComboboxSelect (i) {
@@ -450,12 +460,12 @@ export default {
 
     showComboboxList () {
       if (this.showList != null) { return; }
-      this.$refs.comboboxWithPopover.showComboboxList();
+      this.$refs.comboboxWithPopover?.showComboboxList();
     },
 
     closeComboboxList () {
       if (this.showList != null) { return; }
-      this.$refs.comboboxWithPopover.closeComboboxList();
+      this.$refs.comboboxWithPopover?.closeComboboxList();
     },
 
     getChipButtons () {
@@ -511,20 +521,20 @@ export default {
 
     moveFromInputToChip () {
       this.getLastChipButton().focus();
-      this.$refs.input.blur();
+      this.$refs.input?.blur();
       this.closeComboboxList();
     },
 
     moveFromChipToInput () {
       this.getLastChipButton().blur();
-      this.$refs.input.focus();
+      this.$refs.input?.focus();
       this.showComboboxList();
     },
 
     navigateBetweenChips (target, toLeft) {
       const from = this.getChipButtons().indexOf(target);
       const to = toLeft ? from - 1 : from + 1;
-      if (to < 0 || to >= this.$refs.chips.length) {
+      if (to < 0 || to >= this.$refs.chips?.length) {
         return;
       }
       this.getChipButtons()[from].blur();