+++ recipes/leftbar/group_row/group_row.vue
@@ -7,63 +7,32 @@
     :unread-count-tooltip="unreadCountTooltip"
     :selected="selected"
     :is-typing="isTyping"
+    v-bind="$attrs"
   >
     <template #left>
-      <dt-avatar
-        :initials="avatarInitials"
-        :group="groupCount"
-      >
-        <img
-          v-if="avatarSrc"
-          data-qa="dt-avatar-image"
-          :src="avatarSrc"
-          :alt="avatarInitials"
-        >
-      </dt-avatar>
+      <dt-icon
+        name="users"
+        size="300"
+      />
     </template>
   </dt-recipe-general-row>
 </template>
 
 <script>
 import { DtRecipeGeneralRow } from '@/recipes/leftbar/general_row';
-import DtAvatar from '@/components/avatar/avatar.vue';
+import DtIcon from '@/components/icon/icon.vue';
 
 export default {
   name: 'DtRecipeGroupRow',
 
   components: {
-    DtAvatar,
+    DtIcon,
     DtRecipeGeneralRow,
   },
 
+  inheritAttrs: false,
+
   props: {
-    /**
-     * Optional avatar image url.
-     * if provided, it's also required to provide a value in the `avatarInitials` prop to use
-     * in the alt attribute of the avatar.
-     */
-    avatarSrc: {
-      type: String,
-      default: '',
-    },
-
-    /**
-     * Initials to display on the avatar if avatarSrc is not provided or
-     * alt attr if avatarSrc is provided.
-     */
-    avatarInitials: {
-      type: String,
-      default: '',
-      required: true,
-    },
-
-    /**
-     * Number displayed in avatar to count group members
-     */
-    groupCount: {
-      type: Number,
-      required: true,
-    },
 
     /**
      * Screen reader will read out the number of users in the group using this text. Ex: "2 users"